requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
